noir_profiler
Module flamegraph

Structs

BrilligExecutionSample 🔒

CompilationSample 🔒

FoldedStackItem 🔒

InfernoFlamegraphGenerator 🔒

Traits

FlamegraphGenerator 🔒

Sample 🔒

Functions

add_locations_to_folded_stack_items 🔒

find_callsite_labels 🔒

generate_folded_sorted_lines 🔒

location_to_callsite_label 🔒

to_folded_sorted_lines 🔒

Creates a vector of lines in the format that inferno expects from a nested hashmap of stack items The lines have to be sorted in the following way, exploring the graph in a depth-first manner: main 100 main::foo 0 main::foo::bar 200 main::baz 27 main::baz::qux 800